requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
discord.py>=2.0.0
python-dotenv
pytz
httpx
playwright
apscheduler
